Added GUIImage class to assist creating and handling GUIImage entity/component composition
