+++ pesukarhu/ui/UIFactories.h
@@ -51,23 +51,6 @@ namespace pk
             entityID_t checkedStatusImgEntity;
         };
 
-        struct ImgCreationProperties
-        {
-            ConstraintProperties constraintProperties;
-            float width;
-            float height;
-            vec3 color = { 1, 1, 1 };
-            vec3 highlightColor = { 1, 1, 1 };
-            bool useHighlightColor = false;
-            GUIFilterType filter = GUIFilterType::GUI_FILTER_TYPE_NONE;
-            Texture* pTexture = nullptr;
-            vec4 textureCropping = vec4(0, 0, 1, 1);
-        };
-
-
-        entityID_t create_image(ImgCreationProperties creationProperties);
-
-
         entityID_t create_text(
             const std::string& str, const Font& font,
             ConstraintProperties constraintProperties,
